Move decorators to avoid potential circular import
